requirements.txt
streamlit==1.39.0
flask==3.0.3
openai==1.42.0
faiss-cpu==1.8.0.post1
langchain==0.2.14
PyPDF2==3.0.1
python-docx==1.1.2
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3
chardet==5.2.0
numpy==1.26.4
